[REFACTOR]: CategoryRepository 패키지 변경
- board.repository -> category.repository
